refactor(dashboard): extract ThemeToggle from Sidebar
The theme control is a self-contained widget with two presentations of one
choice: a three-way pill when there is room, a single cycling button when
there isn't. It now lives in ThemeToggle.svelte and takes a `compact` prop.

This moves coupling out of Sidebar rather than adding it. Sidebar previously
reached into the widget with four rules to swap the two forms -- two under
.sidebar-collapsed and two in the 768px breakpoint. ThemeToggle now owns both
branches, sitting next to each other so it is obvious they express the same
intent, and neither component needs :global() to reach the other.

Sidebar 435 -> 334 lines.

Not extracted: the individual .theme-btn inside the loop. It is used once,
its styling assumes the pill around it (transparent, borderless, sized as a
segment), and pulling it out would split .theme-btn:focus-visible away from
its siblings and duplicate the focus-ring declaration.
